refactor: Move isRuntimeBot to BotCalculations.ts to prevent circular dependencies
- This also means that the RuntimeBot interface is now in Bot.ts
